Ensure slack-mode is used when creating a buffer
The notifications buffer is created on-the-fly by the
slack-buffer-update-notification function. Since this doesn't first go
through slack-buffer-create, slack-mode is never enabled for this buffer
and so lui-insert doesn't work on it. Using a separate function to
ensure that any new buffer created by slack gets slack-mode enabled
fixes this situation.