Use lui-time-stamp to show the time stamp
This lets us use correct time stamps for messages from the past with
lui's time-stamp showing mechanism.

The `slack-last-ts' property has to be added in the
`lui-pre-output-hook' because in the `lui-post-output-hook' the
`lui-time-stamp-last' will already have changed. If we don't track this
value then editing a message will mess up the time stamp, since it may
have the same one as the latest message. To clarify, this is not a time
stamp as emacs-slack sees it, but a time stamp as printed by lui (so for
example "[02:03]").

The `ts' property is added _again_ over the entire message in the
`lui-post-output-hook' so that the time stamp is also included. If the
time stamp isn't added like this explicitly the time stamp will have no
`ts' text property and that will mess up the editing functions.

I use a dynamic variable `slack-current-message' to communicate which
message we're working on in the hooks. This variable may be nil, since
any place where `lui-insert' is used directly it will not be bound and
probably shouldn't be either since these don't print normal messages in
normal room buffers (think the "load more" message and the notifications
buffer).

When replacing a message the `lui-time-stamp-last' is let-bound to the
value that was saved in the `lui-pre-output-hook' so that the time
stamps are properly inserted (or not) depending on the
`lui-time-stamp-only-when-changed-p' value.