add (helm "3.0") to Package-Requires
Fix #499

It's not clear what version of `helm` is actually required, but since
`helm-slack.el` was introduced in September 2018 and `helm` version
3.0 was released in August 2018, this seems like a reasonable choice.

Note that 99% of all the downloads of helm are from MELPA which
continuously builds the latest commit from github and serves it. So
it's likely many users are using versions of helm newer than 3.0
depending when they last updated their packages.